refactor: PR #92 CodeRabbit 코멘트 수정 (권한 검증, 테스트 정렬, BDD 패턴)
- ManagerWorkspaceRequestController에 @PreAuthorize("hasAnyRole('MANAGER')") 추가
- CancelWorkspaceRequestTest → CancelWorkspaceRequestTests (복수형)
  - 5개 테스트를 @Nested execute 그룹으로 정렬
  - Mockito BDD 패턴 적용 (given/then/willThrow)
- WorkspaceRequestTest → WorkspaceRequestTests (복수형)
  - 4개 cancel 테스트를 @Nested 그룹으로 정렬
