+++ src/main/java/com/dreamteam/alter/adapter/inbound/manager/workspace/controller/ManagerWorkspaceRequestController.java
@@ -3,6 +3,7 @@ package com.dreamteam.alter.adapter.inbound.manager.workspace.controller;
 import java.util.List;
 
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
@@ -30,6 +31,7 @@ import lombok.RequiredArgsConstructor;
 @RestController
 @RequestMapping("/manager/workspace-requests")
 @RequiredArgsConstructor
+@PreAuthorize("hasAnyRole('MANAGER')")
 @Validated
 public class ManagerWorkspaceRequestController implements ManagerWorkspaceRequestControllerSpec {
